fix: break require cycle between effectMmkv and reportError
