+++ apps/mobile/src/utils/clearMmkvStorageAndEmptyAtoms.ts
@@ -1,7 +1,8 @@
 import AsyncStorage from '@react-native-async-storage/async-storage'
 import {CLEAR_STORAGE_KEY} from './atomUtils/atomWithParsedMmkvStorage'
 // TODO: ASYNC_SENTINEL_KEY import is part of temporary MMKV data loss diagnostic. Remove together with the sentinel.
-import {ASYNC_SENTINEL_KEY, storage} from './mmkv/effectMmkv'
+import {ASYNC_SENTINEL_KEY} from './mmkv/detectMmkvDataLoss'
+import {storage} from './mmkv/effectMmkv'
 
 export default function clearMmkvStorageAndEmptyAtoms(): void {
   // set all atoms to defaultValue
