+++ apps/mobile/src/utils/mmkv/effectMmkv.ts
@@ -1,10 +1,6 @@
-import AsyncStorage from '@react-native-async-storage/async-storage'
 import {Either, flow, Schema, type ParseResult} from 'effect'
-import {File, Paths} from 'expo-file-system'
-import {AppState} from 'react-native'
 import {MMKV} from 'react-native-mmkv'
 import {JsonParseError, JsonStringifyError} from '../fpUtils'
-import reportError from '../reportError'
 import {ReadingFromStoreError, ValueNotSet, WritingToStoreError} from './domain'
 
 export interface EffectMmkv {
@@ -116,61 +112,7 @@ function createEffectMmkv(storage: MMKV): EffectMmkv {
   }
 }
 
-// TODO: Temporary diagnostic to detect silent MMKV data wipes (OnErrorDiscard).
-// Remove once the root cause of user data loss is identified.
-const MMKV_SENTINEL_KEY = '__mmkv_data_exists'
-export const ASYNC_SENTINEL_KEY = '__mmkv_was_populated'
-
-function getMmkvFilesDiagnostics(): Record<string, unknown> {
-  const docDir = Paths.document
-  if (!docDir) return {error: 'no document directory'}
-
-  const dataFile = new File(docDir, 'mmkv/mmkv.default')
-  const crcFile = new File(docDir, 'mmkv/mmkv.default.crc')
-
-  return {
-    dataFileExists: dataFile.exists,
-    dataFileSize: dataFile.exists ? dataFile.size : null,
-    crcFileExists: crcFile.exists,
-    crcFileSize: crcFile.exists ? crcFile.size : null,
-  }
-}
-
-function detectMmkvDataLoss(mmkvInstance: MMKV): void {
-  try {
-    const mmkvSentinel = mmkvInstance.getString(MMKV_SENTINEL_KEY)
-
-    void AsyncStorage.getItem(ASYNC_SENTINEL_KEY)
-      .then((asyncSentinel) => {
-        try {
-          if (!mmkvSentinel && asyncSentinel) {
-            const remainingKeyCount = mmkvInstance.getAllKeys().length
-            const fileDiagnostics = getMmkvFilesDiagnostics()
-
-            reportError(
-              'error',
-              new Error(
-                'MMKV data loss detected: data was previously stored but MMKV is now empty'
-              ),
-              {
-                lastPopulatedAt: asyncSentinel,
-                remainingKeyCount,
-                appState: AppState.currentState,
-                ...fileDiagnostics,
-              }
-            )
-          }
-        } catch {}
-
-        mmkvInstance.set(MMKV_SENTINEL_KEY, Date.now().toString())
-        return AsyncStorage.setItem(ASYNC_SENTINEL_KEY, Date.now().toString())
-      })
-      .catch(() => {})
-  } catch {}
-}
-
 const mmkv = new MMKV()
-detectMmkvDataLoss(mmkv)
 if (__DEV__) {
   // @ts-expect-error for debugging purposes
   window.__mmkv = mmkv
